+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+// Place your settings in this file to overwrite default and user settings.
+{
+  "files.associations": {
+    "*.src": "perl"
+  },
+  "editor.wordWrap": "off",
+  "files.encoding": "utf8",
+  "files.autoGuessEncoding": false
+}
